# Faster basename, dirname, extname