--- a/benches/rust_ruby.rs
+++ b/benches/rust_ruby.rs
@@ -14,7 +14,7 @@ fn rust_to_ruby_c_char_convenient(b: &mut Bencher){
   let s: String = "hello".to_string();
 
   // (185ns)
-  b.iter(|| { 
+  b.iter(|| {
     let s = s.clone(); //To imitate .into()
     let s_slice: &str = &s[..];
     CString::new(s_slice).unwrap().into_raw()
@@ -28,7 +28,7 @@ fn rust_to_ruby_c_char(b: &mut Bencher){
   let s = s.as_str();
 
   // USE THIS METHOD!!! (153ns)
-  b.iter(|| { 
+  b.iter(|| {
     CString::new(s).unwrap().into_raw()
   })
 }
@@ -39,7 +39,7 @@ fn ruby_to_rust_string_convenient(b: &mut Bencher){
   let s: *const c_char = CString::new("hello").unwrap().into_raw();
 
   // (46ns)
-  b.iter(|| { 
+  b.iter(|| {
     let c_str = unsafe {
       assert!(!s.is_null());
       CStr::from_ptr(s)
@@ -51,11 +51,11 @@ fn ruby_to_rust_string_convenient(b: &mut Bencher){
 
 // SANCTIONED USAGE
 #[bench]
-fn ruby_to_rust_from_c_char(b: &mut Bencher){
+fn ruby_to_rust_from_c_char_from_utf8(b: &mut Bencher){
   let s: *const c_char = CString::new("hello").unwrap().into_raw();
 
   // USE THIS METHOD!!! (19ns)
-  b.iter(|| { 
+  b.iter(|| {
     let c_str = unsafe {
       if s.is_null() {
         return "";
@@ -64,4 +64,16 @@ fn ruby_to_rust_from_c_char(b: &mut Bencher){
     };
     str::from_utf8(c_str.to_bytes()).unwrap_or("")
   })
+}
+
+
+#[bench]
+fn ruby_to_rust_from_c_char_to_str(b: &mut Bencher) {
+  let s: *const c_char = CString::new("hello").unwrap().into_raw();
+  b.iter(|| {
+    if s.is_null() {
+      return "";
+    }
+    unsafe { CStr::from_ptr(s) }.to_str().unwrap_or("")
+  });
 }