Benches several ways of working with paths at low level.
Changes the methods to use the fastest approach.

```
$ rake pbench
Building extension...
   Compiling libc v0.2.11
   Compiling faster_path v0.0.1 (file:///home/glebm/repos/glebm/faster_path)
Cleaning up build...
Completed build!
Pinch-bench (Pbench) by Daniel P. Clark
--------------------------------------------------------------------------------
64-bit ruby 2.3.1p112 (2016-04-26 revision 54768) [x86_64-linux]
64-bit rustc 1.11.0-nightly (ad7fe6521 2016-06-23)
--------------------------------------------------------------------------------
Performance change for allocate, instead of new, is 80.6%
Performance change for absolute? is 92.4%
Performance change for add_trailing_separator is 54.4%
Performance change for basename is 4.7%
Performance change for chop_basename is 27.8%
Performance change for directory? is 33.8%
Performance change for dirname is -11.8%
Performance change for extname is 36.1%
Performance change for has_trailing_separator? is 49.5%
Performance change for blank? (verses strip.empty?) is -80.7%
Performance change for relative? is 91.9%
Started with run options --seed 5107
```

cargo bench:
```
$ cargo bench
     Running target/release/faster_path-107c566d7e5973d7

running 1 test
test rust_arch_bits::it_is_32_or_64 ... ignored

test result: ok. 0 passed; 0 failed; 1 ignored; 0 measured

     Running target/release/path_parsing-9132e73b45e51362

running 3 tests
test tests::bench_extract_last_path_segment_bytes   ... bench:           8 ns/iter (+/- 0)
test tests::bench_extract_last_path_segment_pointer ... bench:           5 ns/iter (+/- 1)
test tests::bench_extract_last_path_segment_str     ... bench:          26 ns/iter (+/- 0)

test result: ok. 0 passed; 0 failed; 0 ignored; 3 measured

     Running target/release/rust_ruby-737dd477eaa16f6c

running 5 tests
test ruby_to_rust_from_c_char_from_utf8 ... bench:          11 ns/iter (+/- 0)
test ruby_to_rust_from_c_char_to_str    ... bench:          10 ns/iter (+/- 0)
test ruby_to_rust_string_convenient     ... bench:          21 ns/iter (+/- 2)
test rust_to_ruby_c_char                ... bench:          80 ns/iter (+/- 9)
test rust_to_ruby_c_char_convenient     ... bench:          86 ns/iter (+/- 0)

test result: ok. 0 passed; 0 failed; 0 ignored; 5 measured
```