@danielpclark 2016-06-27
Do the `path_parsing.rs` methods guarantee safe bounds?  Perhaps we should declare them as unsafe otherwise.