@glebm 2016-06-27
I think `unsafe` is only to be used when calling these method can result in unsafe behaviour. But calling one of these methods will never panic (no matter what the arguments are), as confirmed by inspection and tests, so I think these are safe.