@danielpclark 2016-06-30
Sorry I took a while getting to this.  I was starting a video series and wanted to show checking out Github pull requests locally.