@glebm 2016-06-30
No rush! This has been a great project so far to learn me some rust :+1: